+++ djng/app_settings.py
@@ -17,7 +17,7 @@ class AppSettings(object):
         return FileSystemStorage(location=os.path.join(media_root, upload_temp))
 
     @property
-    def THUMBNAIL_SIZE(self):
+    def THUMBNAIL_OPTIONS(self):
         """
         Set the size as a 2-tuple for thumbnailed images after uploading them.
         """
@@ -26,7 +26,7 @@ class AppSettings(object):
         size = self._setting('DJNG_THUMBNAIL_SIZE', (200, 200))
         if not (isinstance(size, (list, tuple)) and len(size) == 2 and isinstance(size[0], int) and isinstance(size[1], int)):
             raise ImproperlyConfigured("'DJNG_THUMBNAIL_SIZE' must be a 2-tuple of integers.")
-        return size
+        return {'crop': True, 'size': size}
 
 
 import sys