+++ djng/static/djng/js/django-angular.js
@@ -5,85 +5,63 @@
 // Connect the third party module `ng-file-upload` to django-angular
 var fileuploadModule = angular.module('djng.fileupload', ['ngFileUpload']);
 
-fileuploadModule.controller('FileUploadController', ['$scope', 'Upload', function($scope, Upload) {
-	this.uploadFiles = function(element, attrs, files) {
-		angular.forEach(files, function(file, index) {
-			var data = {}, identifier = 'file:' + index;
-			if (file.$error)
-				return;
-			data[identifier] = file;
-			element.addClass('uploading');
-			Upload.upload({
-				data: data,
-				url: attrs.fileuploadUrl
-			}).then(function(response) {
-				var field = response.data[identifier];
-				element.removeClass('uploading');
-				element.css('background-image', field.url);
-				delete field.url;  // we don't want to send back the whole image
-				angular.extend($scope.$eval(attrs.ngModel), field);
-			}, function(respose) {
-				element.removeClass('uploading');
-				console.error(respose.statusText);
-			});
-		});
+
+fileuploadModule.directive('djngFileuploadUrl', ['Upload', function(Upload) {
+	return {
+		restrict: 'A',
+		require: 'ngModel',
+		link: function(scope, element, attrs, ngModelController) {
+			ngModelController.$setViewValue({});
+			element.data('area_label', element.val());
+			if (attrs.currentFile) {
+				angular.extend(scope.$eval(attrs.ngModel), {current_file: attrs.currentFile});
+				element.data('current_file', attrs.currentFile);
+				element.val(attrs.currentFile.substring(0, attrs.currentFile.indexOf(':')));
+				element.addClass('djng-preset');
+			} else {
+				element.addClass('djng-empty');
+			}
+
+			scope.uploadFile = function(file, filetype, id, model) {
+				var data = {'file:0': file, filetype: filetype},
+				    element = angular.element(document.querySelector('#' + id));
+				element.addClass('uploading');
+				Upload.upload({
+					data: data,
+					url: attrs.djngFileuploadUrl
+				}).then(function(response) {
+					var field = response.data['file:0'], current = element.data('current_file');
+					element.removeClass('uploading');
+					element.css('background-image', field.url);
+					element.removeClass('djng-empty')
+					element.removeClass('djng-preset')
+					element.val(field.file_name);
+					delete field.url;  // we don't want to send back the whole image
+					angular.extend(scope.$eval(model), field, current ? {current_file: current} : {});
+				}, function(respose) {
+					element.removeClass('uploading');
+					console.error(respose.statusText);
+				});
+			};
+		}
 	};
 }]);
 
-
-fileuploadModule.directive('ngfDrop', function() {
-	return {
-		restrict: 'AEC',
-		controller: 'FileUploadController',
-		require: ['ngfDrop', 'ngModel'],
-		link: function(scope, element, attrs, ctrls) {
-			var fileUploadController = ctrls[0], ngModelController = ctrls[1];
-
-			ngModelController.$setViewValue({
-				previous_image: attrs.previousImage,
-				temp_name: attrs.previousImage ? true : null
-			});
-
-			scope.uploadFiles = function(files) {
-				fileUploadController.uploadFiles(element, attrs, files);
-			};
-
-			scope.getClass = function() {
-				var model = ngModelController.$viewValue;
-				if (!model || !model.temp_name)
-					return 'empty';
-			};
-		}
-	};
-});
-
-
-fileuploadModule.directive('ngfSelect', function() {
-	return {
-		restrict: 'AEC',
-		controller: 'FileUploadController',
-		link: function(scope, element, attrs, fileUploadController) {
-			scope.uploadFiles = function(files) {
-				fileUploadController.uploadFiles(element, attrs, files);
-			};
-		}
-	};
-});
 
 fileuploadModule.directive('djngFileuploadButton', function() {
 	return {
 		restrict: 'A',
 		link: function(scope, element, attrs) {
-			scope.deleteImage = function() {
-				var model = scope.$eval(attrs.djngFileuploadButton);
+			scope.deleteImage = function(id, model) {
+				var model = scope.$eval(model),
+				    element = angular.element(document.querySelector('#' + id));
+				element.css('background-image', 'none');
+				element.addClass('djng-empty');
+				element.removeClass('djng-preset');
+				element.val(element.data('area_label'));
 				if (model) {
-					model.temp_name = false;  // tags previous image for deletion
+					model.temp_name = 'delete';  // tags previous image for deletion
 				}
-			};
-
-			scope.isEmpty = function() {
-				var model = scope.$eval(attrs.djngFileuploadButton);
-				return !(model && model.temp_name);
 			};
 		}
 	};