+++ djng/styling/bootstrap3/forms.py
@@ -1,9 +1,17 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
+
+import warnings
+
+from django import VERSION as DJANGO_VERSION
 from django.forms.forms import BaseForm
 from django.forms.models import BaseModelForm
 from django.utils import six
 from djng.forms import NgDeclarativeFieldsMetaclass, NgModelFormMetaclass, NgFormBaseMixin
+
+
+if DJANGO_VERSION >= (1, 11):
+    warnings.warn("Since Django-1.11 `djng.styling.bootstrap3` is deprecated.", PendingDeprecationWarning)
 
 
 class Bootstrap3FormMixin(object):
@@ -11,7 +19,7 @@ class Bootstrap3FormMixin(object):
     widget_css_classes = 'form-control'
     form_error_css_classes = 'djng-form-errors'
     field_error_css_classes = 'djng-form-control-feedback djng-field-errors'
-    field_mixins_module = 'djng.styling.bootstrap3.field_mixins'
+    widgets_module = 'djng.styling.bootstrap3.widgets'
     label_css_classes = 'control-label'
 
     def as_div(self):