+++ examples/server/settings.py
@@ -74,6 +74,8 @@ STATICFILES_DIRS = (
     ('node_modules', os.environ.get('NODE_MODULES_DIR', os.path.join(APP_DIR, 'examples/node_modules'))),
 )
 
+FORM_RENDERER = 'djng.forms.renderers.DjangoAngularBootstrap3Templates'
+
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',