+++ examples/server/tests/settings.py
@@ -55,6 +55,8 @@ STATICFILES_DIRS = (
     os.path.join(BASE_DIR, 'client', 'src'),
 )
 
+FORM_RENDERER = 'djng.forms.renderers.DjangoAngularTemplates'
+
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',