+++ examples/server/tests/test_fileupload.py
@@ -7,19 +7,17 @@ from django.conf import settings
 from django.core.urlresolvers import reverse
 from django.core import signing
 from django.core.files.uploadedfile import InMemoryUploadedFile, TemporaryUploadedFile
-from django.db.models.fields.files import ImageFieldFile
-from django.forms import forms
 from django.test import override_settings, TestCase
 from django.test.client import Client
 
 from pyquery.pyquery import PyQuery
 
 from djng import app_settings
-from djng.forms import NgModelFormMixin
+from djng.forms import NgModelFormMixin, NgForm
 from djng.forms.fields import ImageField
 
 
-class TestUploadForm(NgModelFormMixin, forms.Form):
+class TestUploadForm(NgModelFormMixin, NgForm):
     scope_prefix = 'my_data'
     form_name = 'my_form'
 
@@ -42,7 +40,7 @@ class FileUploadTest(TestCase):
         upload_filename = os.path.join(os.path.dirname(__file__), 'sample-image.jpg')
         with open(upload_filename, 'rb') as fp:
             upload_url = reverse('fileupload')
-            response = client.post(upload_url, {'file:0': fp})
+            response = client.post(upload_url, {'file:0': fp, 'filetype': 'image'})
         self.assertEquals(response.status_code, 200)
         return json.loads(response.content.decode('utf-8'))
 
@@ -50,7 +48,7 @@ class FileUploadTest(TestCase):
         content = self.upload_image()
         self.assertTrue('file:0' in content)
         self.assertTrue(content['file:0']['url'].startswith('url(data:application/octet-stream;base64,/9j/4AAQSkZJRgABA'))
-        self.assertEquals(content['file:0']['image_name'], 'sample-image.jpg')
+        self.assertEquals(content['file:0']['file_name'], 'sample-image.jpg')
         self.assertEquals(content['file:0']['content_type'], 'image/jpeg')
         self.assertEquals(self.signer.unsign(content['file:0']['temp_name']), 'sample-image.jpg')
 
@@ -59,13 +57,14 @@ class FileUploadTest(TestCase):
         htmlsource = form.as_p()
         dom = PyQuery(htmlsource)
         textarea = dom('div.drop-box textarea')
-        self.assertEquals(textarea.attr('fileupload-url'), reverse('fileupload'))
-        self.assertEquals(textarea.attr('ngf-drop'), 'uploadFiles($files)')
-        self.assertEquals(textarea.attr('ngf-select'), 'uploadFiles($files)')
+        self.assertEquals(textarea.attr('djng-fileupload-url'), reverse('fileupload'))
+        self.assertEquals(textarea.attr('ngf-drop'), 'uploadFile($file, "image", "id_avatar", "my_data[\'avatar\']")')
+        self.assertEquals(textarea.attr('ngf-select'), 'uploadFile($file, "image", "id_avatar", "my_data[\'avatar\']")')
 
-        delete_button = dom('div.drop-box span')
-        self.assertEquals(delete_button.attr('djng-fileupload-button'), "my_data['avatar']")
-        self.assertEquals(delete_button.attr('ng-click'), "deleteImage()")
+        delete_button = dom('div.drop-box img.djng-btn-trash')
+        self.assertEquals(delete_button.attr('src'), '/static/djng/icons/image/trash.svg')
+        self.assertEquals(delete_button.attr('djng-fileupload-button'), '')
+        self.assertEquals(delete_button.attr('ng-click'), 'deleteImage("id_avatar", "my_data[\'avatar\']")')
 
     def test_receive_small_image(self):
         content = self.upload_image()